requirements.txt
protobuf==4.21.7
grpcio==1.49.1
grpcio-tools==1.49.1
